cnnmodels/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.196
Pillow==10.0.1
numpy==1.24.3
pydantic==2.5.0
